APP-761: Initial version of PullRequest parser v2.
